added exceptionDTO for exception handling, added limit logic, user role can be successfully updated

+++ IT2B-G05-BankApp/it2bank/src/main/java/nl/inholland/it2bank/service/UserService.java
@@ -47,7 +47,7 @@ public class UserService {
         user.setEmail(userDto.email());
         user.setPassword(userDto.password());
         user.setPhoneNumber(userDto.phoneNumber());
-        user.setRole(userDto.role());
+        user.setRole(UserRoles.valueOf(userDto.role()));
         user.setTransactionLimit(userDto.transactionLimit());
         user.setDailyLimit(userDto.dailyLimit());
 
@@ -59,12 +59,37 @@ public class UserService {
                 .orElseThrow(() -> new EntityNotFoundException("User not found"));
     }
 
-    public void deleteUser(long id){
-        userRepository.deleteById(id);
+    public void updateUser(long id, UserDTO updatedUser){
+        UserModel existingUser = this.getUserById(id);
+
+        if(updatedUser.transactionLimit() > updatedUser.dailyLimit()){
+            throw new IllegalArgumentException("Transaction limit cannot be higher than daily limit.");
+        }else{
+            existingUser.setFirstName(updatedUser.firstName());
+            existingUser.setLastName(updatedUser.lastName());
+            existingUser.setBsn(updatedUser.bsn());
+            existingUser.setEmail(updatedUser.email());
+            existingUser.setPhoneNumber(updatedUser.phoneNumber());
+            existingUser.setPassword(updatedUser.password());
+            existingUser.setRole(UserRoles.valueOf(updatedUser.role()));
+            existingUser.setDailyLimit(updatedUser.dailyLimit());
+            existingUser.setTransactionLimit(updatedUser.transactionLimit());
+
+            userRepository.save(existingUser);
+        }
     }
 
-    public UserModel saveUser(UserModel user){
-        return userRepository.save(user);
+    public void deleteUser(long id){
+        UserModel user = this.getUserById(id);
+        if(user.getRole() == UserRoles.User) {
+            userRepository.deleteById(id);
+        }else{
+            throw new IllegalArgumentException("User cannot be deleted");
+        }
+    }
+
+    public UserModel saveUser(UserDTO user){
+        return userRepository.save(this.mapObjectToUser(user));
     }
 
     public String login(String email, String password) throws Exception {
